agent-service/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Muhammad Ahmad <ahmadshaukat_4@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = "^0.115.0"
uvicorn = "^0.30.3"
langchain = "^0.3.1"
langchain-core = "^0.3.6"
openai = "^1.50.2"
langchain-community = "^0.3.1"
langgraph = "^0.2.28"
langserve = {extras = ["all"], version = "^0.3.0"}
pydantic = "^2.9.2"
langchain-openai = "^0.2.1"
langchain-groq = "^0.2.0"
langchain-google-genai = "^2.0.0"
langgraph-checkpoint = "^2.0.0"
duckduckgo-search = "^6.2.13"
pyowm = "^3.3.0"
httpx = "^0.27.2"
langsmith = "^0.1.129"
python-dotenv = "^1.0.1"
setuptools = "^75.1.0"
tavily-python = "^0.5.0"
pandas = "^2.2.3"
ipython = "^8.28.0"
pytest = "^8.3.3"
psycopg = "^3.2.3"
psycopg-pool = "^3.2.3"
langgraph-checkpoint-postgres = "^2.0.1"
redis = "^5.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
